Extract shared Cranelift function compile path
