+++ crates/stasis_compiler/src/backend/emit.rs
@@ -13,9 +13,8 @@ use cranelift_codegen::ir::{
     immediates::{Ieee32, Ieee64},
     types, AbiParam, Block, FuncRef, InstBuilder, MemFlags, Value,
 };
-use cranelift_frontend::{FunctionBuilder, Variable};
+use cranelift_frontend::{FunctionBuilder, FunctionBuilderContext, Variable};
 use cranelift_module::{FuncId, Linkage, Module};
-use cranelift_object::ObjectModule;
 use std::collections::{BTreeMap, HashMap};
 use std::path::{Path, PathBuf};
 
@@ -260,8 +259,7 @@ pub(crate) fn resolve_preferred_extern_call_signatures(
     extern_signatures: &[ExternCallSignature],
 ) -> Result<(Vec<ResolvedExternCallSignature>, ExternSymbolAddressMap), String> {
     resolve_extern_call_signatures_with(extern_signatures, |signature, candidate| {
-        if is_known_aot_runtime_extern_symbol(candidate) || signature.symbol_candidates.len() == 1
-        {
+        if is_known_aot_runtime_extern_symbol(candidate) || signature.symbol_candidates.len() == 1 {
             Some(0)
         } else {
             None
@@ -276,7 +274,10 @@ pub(crate) fn build_compile_analysis_cache(
     files_fingerprint: u64,
     extern_resolver: impl FnOnce(
         &[ExternCallSignature],
-    ) -> Result<(Vec<ResolvedExternCallSignature>, ExternSymbolAddressMap), String>,
+    ) -> Result<
+        (Vec<ResolvedExternCallSignature>, ExternSymbolAddressMap),
+        String,
+    >,
 ) -> Result<CompileAnalysisCache, String> {
     let extern_signatures = collect_supported_extern_call_signatures(files, type_table)?;
     let (resolved_extern_signatures, extern_symbol_addresses) =
@@ -1339,7 +1340,7 @@ pub(crate) struct RuntimeCallRefs {
 }
 
 pub(crate) struct AotDirectCallMode<'a> {
-    pub(crate) module: &'a mut ObjectModule,
+    pub(crate) module: &'a mut dyn Module,
     pub(crate) self_function_id: FunctionId,
     pub(crate) self_clif_func_id: FuncId,
     pub(crate) imported_function_ids: HashMap<FunctionId, FuncId>,
@@ -1348,6 +1349,12 @@ pub(crate) struct AotDirectCallMode<'a> {
 pub(crate) enum InternalCallMode<'a> {
     Jit,
     AotDirect(AotDirectCallMode<'a>),
+}
+
+#[derive(Clone, Copy, Debug, PartialEq, Eq)]
+pub(crate) enum SharedCompileBackendMode {
+    Jit,
+    AotDirect,
 }
 
 fn aot_symbol_name(function_id: FunctionId) -> String {
@@ -1441,6 +1448,247 @@ pub(crate) struct LocalBinding {
 pub(crate) struct StructViewBinding {
     pub(crate) index_var: Variable,
     pub(crate) len_var: Variable,
+}
+
+pub(crate) fn compile_function_with_module<M, T, BeforeStatement, OnFunctionBuilt, Finalize>(
+    mut module: M,
+    meta: &FunctionMeta,
+    hir: &FunctionHIR,
+    symbol: &str,
+    backend_mode: SharedCompileBackendMode,
+    call_signatures: &CallSignatureMap,
+    type_table: &mut TypeTable,
+    global_path_types: &GlobalPathTypeMap,
+    constant_values: &ConstantValueMap,
+    collection_infos: &CollectionInfoMap,
+    named_struct_field_types: &NamedStructFieldTypeMap,
+    mut before_statement: BeforeStatement,
+    on_function_built: OnFunctionBuilt,
+    finalize: Finalize,
+) -> Result<T, String>
+where
+    M: Module,
+    BeforeStatement: FnMut(&SimpleStmt) -> Result<(), String>,
+    OnFunctionBuilt: FnOnce(&FunctionMeta, &cranelift_codegen::ir::Function),
+    Finalize: FnOnce(M, FuncId, cranelift_codegen::Context) -> Result<T, String>,
+{
+    let mut context = module.make_context();
+    context.func.signature = module.make_signature();
+    for param_type in &meta.params {
+        append_abi_params_for_type_id(
+            &mut context.func.signature.params,
+            *param_type,
+            type_table,
+            named_struct_field_types,
+        )?;
+    }
+    if meta.return_type != TYPE_ID_VOID {
+        let clif_return_type =
+            clif_type_for_type_id(meta.return_type, type_table).map_err(|_| {
+                format!(
+                    "unsupported return type id {} for function {}",
+                    meta.return_type, meta.name
+                )
+            })?;
+        context
+            .func
+            .signature
+            .returns
+            .push(AbiParam::new(clif_return_type));
+    }
+
+    let function_id = module
+        .declare_function(symbol, Linkage::Export, &context.func.signature)
+        .map_err(|error| format!("failed to declare function {symbol}: {error}"))?;
+    let runtime_call_imports = build_runtime_call_import_ids(
+        &mut module,
+        call_signatures,
+        type_table,
+        named_struct_field_types,
+    )?;
+
+    let mut function_builder_context = FunctionBuilderContext::new();
+    {
+        let mut builder = FunctionBuilder::new(&mut context.func, &mut function_builder_context);
+        let runtime_call_refs =
+            build_runtime_call_refs(&mut module, &runtime_call_imports, builder.func);
+        let entry = builder.create_block();
+        for param_type in &meta.params {
+            if is_struct_view_type(*param_type, named_struct_field_types) {
+                for _ in 0..STRUCT_VIEW_ABI_WORDS {
+                    builder.append_block_param(entry, types::I32);
+                }
+            } else {
+                builder.append_block_param(entry, clif_type_for_type_id(*param_type, type_table)?);
+            }
+        }
+        builder.switch_to_block(entry);
+        builder.seal_block(entry);
+
+        if meta.param_names.len() != meta.params.len() {
+            return Err(format!(
+                "parameter metadata mismatch for function '{}' ({} names, {} types)",
+                meta.name,
+                meta.param_names.len(),
+                meta.params.len()
+            ));
+        }
+        let mut values_by_name: BTreeMap<String, LocalBinding> = BTreeMap::new();
+        let mut next_variable = 0u32;
+        let block_params: Vec<Value> = builder.block_params(entry).to_vec();
+        let mut block_param_cursor = 0usize;
+        for (index, name) in meta.param_names.iter().enumerate() {
+            let param_type = meta.params[index];
+            let (variable, struct_view) =
+                if is_struct_view_type(param_type, named_struct_field_types) {
+                    let base_value =
+                        block_params
+                            .get(block_param_cursor)
+                            .copied()
+                            .ok_or_else(|| {
+                                format!(
+                                    "missing struct view base parameter {} for function '{}'",
+                                    block_param_cursor, meta.name
+                                )
+                            })?;
+                    let index_value = block_params
+                        .get(block_param_cursor + 1)
+                        .copied()
+                        .ok_or_else(|| {
+                            format!(
+                                "missing struct view index parameter {} for function '{}'",
+                                block_param_cursor + 1,
+                                meta.name
+                            )
+                        })?;
+                    let len_value = block_params
+                        .get(block_param_cursor + 2)
+                        .copied()
+                        .ok_or_else(|| {
+                            format!(
+                                "missing struct view len parameter {} for function '{}'",
+                                block_param_cursor + 2,
+                                meta.name
+                            )
+                        })?;
+                    block_param_cursor += STRUCT_VIEW_ABI_WORDS;
+
+                    let base_var = declare_new_variable(
+                        &mut builder,
+                        &mut next_variable,
+                        base_value,
+                        param_type,
+                        type_table,
+                    )?;
+                    let index_var = declare_new_variable(
+                        &mut builder,
+                        &mut next_variable,
+                        index_value,
+                        TYPE_ID_I32,
+                        type_table,
+                    )?;
+                    let len_var = declare_new_variable(
+                        &mut builder,
+                        &mut next_variable,
+                        len_value,
+                        TYPE_ID_I32,
+                        type_table,
+                    )?;
+                    (base_var, Some(StructViewBinding { index_var, len_var }))
+                } else {
+                    let value = block_params
+                        .get(block_param_cursor)
+                        .copied()
+                        .ok_or_else(|| {
+                            format!(
+                                "missing block parameter {} for function '{}'",
+                                block_param_cursor, meta.name
+                            )
+                        })?;
+                    block_param_cursor = block_param_cursor.saturating_add(1);
+                    (
+                        declare_new_variable(
+                            &mut builder,
+                            &mut next_variable,
+                            value,
+                            param_type,
+                            type_table,
+                        )?,
+                        None,
+                    )
+                };
+            if values_by_name.contains_key(name) {
+                return Err(format!("parameter '{}' shadows existing variable", name));
+            }
+            values_by_name.insert(
+                name.clone(),
+                LocalBinding {
+                    var: variable,
+                    type_id: param_type,
+                    struct_view,
+                },
+            );
+        }
+        if block_param_cursor != block_params.len() {
+            return Err(format!(
+                "block parameter count mismatch for function '{}' (consumed {}, found {})",
+                meta.name,
+                block_param_cursor,
+                block_params.len()
+            ));
+        }
+
+        let empty_foreach_bindings = ForeachBindingMap::new();
+        let mut internal_calls = match backend_mode {
+            SharedCompileBackendMode::Jit => InternalCallMode::Jit,
+            SharedCompileBackendMode::AotDirect => InternalCallMode::AotDirect(AotDirectCallMode {
+                module: &mut module,
+                self_function_id: meta.id,
+                self_clif_func_id: function_id,
+                imported_function_ids: HashMap::new(),
+            }),
+        };
+        let body = extract_function_body(hir)?;
+        let mut terminated = false;
+        parse_simple_statements_from_block_with(body, type_table, |type_table, statement| {
+            if terminated {
+                return Ok(());
+            }
+            before_statement(&statement)?;
+            terminated = emit_simple_statements(
+                &mut builder,
+                std::slice::from_ref(&statement),
+                &mut values_by_name,
+                &runtime_call_refs,
+                &mut internal_calls,
+                call_signatures,
+                type_table,
+                global_path_types,
+                constant_values,
+                collection_infos,
+                named_struct_field_types,
+                &empty_foreach_bindings,
+                None,
+                meta.return_type,
+                &mut next_variable,
+            )?;
+            Ok(())
+        })?;
+        if !terminated {
+            if meta.return_type == TYPE_ID_VOID {
+                builder.ins().return_(&[]);
+            } else {
+                return Err(format!(
+                    "non-void function '{}' must end with a return statement",
+                    meta.name
+                ));
+            }
+        }
+        builder.finalize();
+    }
+
+    on_function_built(meta, &context.func);
+    finalize(module, function_id, context)
 }
 
 #[derive(Clone, Copy, Debug, PartialEq, Eq)]
