+++ crates/stasis_compiler/src/backend/jit.rs
@@ -5,10 +5,8 @@ use crate::frontend::types::{
     TypeCategory, TypeTable, TYPE_ID_BOOL, TYPE_ID_F32, TYPE_ID_I32, TYPE_ID_VOID,
 };
 use crate::ir::hir::FunctionHIR;
-use cranelift_codegen::ir::{types, AbiParam, InstBuilder, Value};
-use cranelift_frontend::{FunctionBuilder, FunctionBuilderContext};
 use cranelift_jit::{JITBuilder, JITModule};
-use cranelift_module::{default_libcall_names, Linkage, Module};
+use cranelift_module::{default_libcall_names, Module};
 use std::collections::{BTreeMap, BTreeSet, HashMap};
 use std::path::Path;
 #[cfg(test)]
@@ -956,222 +954,32 @@ fn compile_function_to_jit_module(
         }
         jit_builder.symbol(extern_symbol, *address as *const u8);
     }
-    let mut module = JITModule::new(jit_builder);
-    let mut context = module.make_context();
-    context.func.signature = module.make_signature();
-    for param_type in &meta.params {
-        append_abi_params_for_type_id(
-            &mut context.func.signature.params,
-            *param_type,
-            type_table,
-            named_struct_field_types,
-        )?;
-    }
-    if meta.return_type != TYPE_ID_VOID {
-        let clif_return_type =
-            clif_type_for_type_id(meta.return_type, type_table).map_err(|_| {
-                format!(
-                    "unsupported JIT return type id {} for function {}",
-                    meta.return_type, meta.name
-                )
-            })?;
-        context
-            .func
-            .signature
-            .returns
-            .push(AbiParam::new(clif_return_type));
-    }
-
-    let function_id = module
-        .declare_function(symbol, Linkage::Export, &context.func.signature)
-        .map_err(|error| format!("failed to declare JIT function {symbol}: {error}"))?;
-    let runtime_call_imports = build_runtime_call_import_ids(
-        &mut module,
+    compile_function_with_module(
+        JITModule::new(jit_builder),
+        meta,
+        hir,
+        symbol,
+        SharedCompileBackendMode::Jit,
         call_signatures,
         type_table,
+        global_path_types,
+        constant_values,
+        collection_infos,
         named_struct_field_types,
-    )?;
-
-    let mut function_builder_context = FunctionBuilderContext::new();
-    {
-        let mut builder = FunctionBuilder::new(&mut context.func, &mut function_builder_context);
-        let runtime_call_refs =
-            build_runtime_call_refs(&mut module, &runtime_call_imports, builder.func);
-        let entry = builder.create_block();
-        for param_type in &meta.params {
-            if is_struct_view_type(*param_type, named_struct_field_types) {
-                for _ in 0..STRUCT_VIEW_ABI_WORDS {
-                    builder.append_block_param(entry, types::I32);
-                }
-            } else {
-                builder.append_block_param(entry, clif_type_for_type_id(*param_type, type_table)?);
-            }
-        }
-        builder.switch_to_block(entry);
-        builder.seal_block(entry);
-
-        if meta.param_names.len() != meta.params.len() {
-            return Err(format!(
-                "parameter metadata mismatch for function '{}' ({} names, {} types)",
-                meta.name,
-                meta.param_names.len(),
-                meta.params.len()
-            ));
-        }
-        let mut values_by_name: BTreeMap<String, LocalBinding> = BTreeMap::new();
-        let mut next_variable = 0u32;
-        let block_params: Vec<Value> = builder.block_params(entry).to_vec();
-        let mut block_param_cursor = 0usize;
-        for (index, name) in meta.param_names.iter().enumerate() {
-            let param_type = meta.params[index];
-            let (variable, struct_view) =
-                if is_struct_view_type(param_type, named_struct_field_types) {
-                    let base_value =
-                        block_params
-                            .get(block_param_cursor)
-                            .copied()
-                            .ok_or_else(|| {
-                                format!(
-                                    "missing struct view base parameter {} for function '{}'",
-                                    block_param_cursor, meta.name
-                                )
-                            })?;
-                    let index_value = block_params
-                        .get(block_param_cursor + 1)
-                        .copied()
-                        .ok_or_else(|| {
-                            format!(
-                                "missing struct view index parameter {} for function '{}'",
-                                block_param_cursor + 1,
-                                meta.name
-                            )
-                        })?;
-                    let len_value = block_params
-                        .get(block_param_cursor + 2)
-                        .copied()
-                        .ok_or_else(|| {
-                            format!(
-                                "missing struct view len parameter {} for function '{}'",
-                                block_param_cursor + 2,
-                                meta.name
-                            )
-                        })?;
-                    block_param_cursor += STRUCT_VIEW_ABI_WORDS;
-
-                    let base_var = declare_new_variable(
-                        &mut builder,
-                        &mut next_variable,
-                        base_value,
-                        param_type,
-                        type_table,
-                    )?;
-                    let index_var = declare_new_variable(
-                        &mut builder,
-                        &mut next_variable,
-                        index_value,
-                        TYPE_ID_I32,
-                        type_table,
-                    )?;
-                    let len_var = declare_new_variable(
-                        &mut builder,
-                        &mut next_variable,
-                        len_value,
-                        TYPE_ID_I32,
-                        type_table,
-                    )?;
-                    (base_var, Some(StructViewBinding { index_var, len_var }))
-                } else {
-                    let value = block_params
-                        .get(block_param_cursor)
-                        .copied()
-                        .ok_or_else(|| {
-                            format!(
-                                "missing block parameter {} for function '{}'",
-                                block_param_cursor, meta.name
-                            )
-                        })?;
-                    block_param_cursor = block_param_cursor.saturating_add(1);
-                    (
-                        declare_new_variable(
-                            &mut builder,
-                            &mut next_variable,
-                            value,
-                            param_type,
-                            type_table,
-                        )?,
-                        None,
-                    )
-                };
-            if values_by_name.contains_key(name) {
-                return Err(format!("parameter '{}' shadows existing variable", name));
-            }
-            values_by_name.insert(
-                name.clone(),
-                LocalBinding {
-                    var: variable,
-                    type_id: param_type,
-                    struct_view,
-                },
-            );
-        }
-        if block_param_cursor != block_params.len() {
-            return Err(format!(
-                "block parameter count mismatch for function '{}' (consumed {}, found {})",
-                meta.name,
-                block_param_cursor,
-                block_params.len()
-            ));
-        }
-
-        let empty_foreach_bindings = ForeachBindingMap::new();
-        let mut internal_calls = InternalCallMode::Jit;
-        let body = extract_function_body(hir)?;
-        let mut terminated = false;
-        parse_simple_statements_from_block_with(body, type_table, |type_table, statement| {
-            if terminated {
-                return Ok(());
-            }
-            terminated = emit_simple_statements(
-                &mut builder,
-                std::slice::from_ref(&statement),
-                &mut values_by_name,
-                &runtime_call_refs,
-                &mut internal_calls,
-                call_signatures,
-                type_table,
-                global_path_types,
-                constant_values,
-                collection_infos,
-                named_struct_field_types,
-                &empty_foreach_bindings,
-                None,
-                meta.return_type,
-                &mut next_variable,
-            )?;
-            Ok(())
-        })?;
-        if !terminated {
-            if meta.return_type == TYPE_ID_VOID {
-                builder.ins().return_(&[]);
-            } else {
-                return Err(format!(
-                    "non-void function '{}' must end with a return statement",
-                    meta.name
-                ));
-            }
-        }
-        builder.finalize();
-    }
-
-    module
-        .define_function(function_id, &mut context)
-        .map_err(|error| format!("failed to define JIT function {symbol}: {error}"))?;
-    module.clear_context(&mut context);
-    module
-        .finalize_definitions()
-        .map_err(|error| format!("failed to finalize JIT definitions: {error}"))?;
-    let code_ptr = module.get_finalized_function(function_id) as usize as u64;
-    Ok((module, code_ptr))
+        |_| Ok(()),
+        |_, _| {},
+        |mut module, function_id, mut context| {
+            module
+                .define_function(function_id, &mut context)
+                .map_err(|error| format!("failed to define JIT function {symbol}: {error}"))?;
+            module.clear_context(&mut context);
+            module
+                .finalize_definitions()
+                .map_err(|error| format!("failed to finalize JIT definitions: {error}"))?;
+            let code_ptr = module.get_finalized_function(function_id) as usize as u64;
+            Ok((module, code_ptr))
+        },
+    )
 }
 
 #[cfg(test)]
